Address PR comments and move modal outside of card
